task 0 of structure and type def project

+++ 0x0E-structures_typedef/dog.h
@@ -4,7 +4,7 @@
 struct dog
 {
 	char *name;
-	double age;
+	float age;
 	char *owner;
 };
